Completed signup confirm logic

+++ src/domain/permission.cr
@@ -6,7 +6,7 @@ class Permission
   include DB::Serializable::NonStrict
   include JSON::Serializable
 
-  property id : Int64
+  property id : Int32
   property name : String
   property is_active = true
 end
