+++ src/domain/song_in_list.cr
@@ -6,7 +6,7 @@ class SongInList
   include DB::Serializable::NonStrict
   include JSON::Serializable
 
-  property id : Int64
+  property id : Int32
   property position : Int32
   property song_id : String
   property song_list_id : String
