+++ src/domain/song_list.cr
@@ -9,5 +9,5 @@ class SongList
   property id : String
   property name : String
   property created_at : Time
-  property user_id : Int64
+  property user_id : Int32
 end
